Fix bug in UnsafeRowIterator when used empty. Fix reported warning in -Wall Improve documentation Improve progress output while processing backup Speed up restoring process (metadata cache is persistent between files) Fix a bug in the BSC compression library usage that prevented maximum compression to be attained

+++ ClassPath/src/Streams/Streams.cpp
@@ -256,7 +256,7 @@ namespace Stream
     bool AESInputStream::goForward(const uint64 skipAmount)
     {
         // Need to read the skip amount multiple time until we're done.
-        if (skipAmount <= (keySize - tempPos))
+        if (skipAmount <= (uint64)(keySize - tempPos))
         {
             tempPos += (uint16)min(skipAmount, (uint64)-1);
             return true;
